fix: legacy layout footer
