+++ frontend/src/lib/components/common/Footer.svelte
@@ -1,16 +1,11 @@
 <footer>
-  <slot />
+    <slot />
 </footer>
 
 <style lang="scss">
   @use "@dfinity/gix-components/styles/mixins/media";
 
   footer {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-
     height: var(--footer-height);
 
     will-change: transform;
