+++ frontend/src/lib/components/common/Layout.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
   import Banner from "../header/Banner.svelte";
   import MenuItems from "./MenuItems.svelte";
-  import {
-    layoutTitleStore,
-    layoutBackStore,
-    layoutMainStyleStore,
-  } from "../../stores/layout.store";
+  import { layoutTitleStore, layoutBackStore } from "../../stores/layout.store";
   import { Layout } from "@dfinity/gix-components";
   import AccountMenu from "../header/AccountMenu.svelte";
   import { triggerDebugReport } from "../../services/debug.services";
@@ -16,11 +12,7 @@
 
 <Banner />
 
-<Layout
-  {back}
-  modern={$layoutMainStyleStore === "modern"}
-  on:nnsBack={() => $layoutBackStore?.()}
->
+<Layout {back} on:nnsBack={() => $layoutBackStore?.()}>
   <h4 use:triggerDebugReport slot="title">{$layoutTitleStore}</h4>
 
   <MenuItems slot="menu-items" />
