+++ frontend/src/routes/Neurons.svelte
@@ -3,23 +3,31 @@
   import { ENABLE_SNS } from "../lib/constants/environment.constants";
   import NnsNeurons from "../lib/pages/NnsNeurons.svelte";
   import SnsNeurons from "../lib/pages/SnsNeurons.svelte";
+  import NnsNeuronsFooter from "../lib/components/neurons/NnsNeuronsFooter.svelte";
   import {
     isNnsProjectStore,
     snsProjectSelectedStore,
   } from "../lib/stores/projects.store";
 </script>
 
-{#if ENABLE_SNS}
-  <div class="dropdown-wrapper">
-    <div class="fit-content">
-      <SelectProjectDropdown />
+<main class="legacy">
+  {#if ENABLE_SNS}
+    <div class="dropdown-wrapper">
+      <div class="fit-content">
+        <SelectProjectDropdown />
+      </div>
     </div>
-  </div>
-{/if}
+  {/if}
+
+  {#if $isNnsProjectStore}
+    <NnsNeurons />
+  {:else if $snsProjectSelectedStore !== undefined}
+    <SnsNeurons />
+  {/if}
+</main>
+
 {#if $isNnsProjectStore}
-  <NnsNeurons />
-{:else if $snsProjectSelectedStore !== undefined}
-  <SnsNeurons />
+  <NnsNeuronsFooter />
 {/if}
 
 <style lang="scss">
